Reordering boxShadow values (#225)
Moved the blur key to between y and spread. If the edit token modal is pulling this index order to generate the fields, it will allow the user to input values in the same sequence as CSS syntax of "box-shadow: h-offset v-offset blur spread color;". This will also match the tab order of Figma's drop shadow values.